feat: migrate existing data consortia roadmap to next.js (#2679)
