--- content/overview/data-consortia.md
@@ -1,52 +1,0 @@
----
-author: "AnVIL"
-description: "AnVIL hosts high value genomic datasets relevant to human health and disease."
-title: "Data Consortia"
----
-
-# Data Consortia
-
-<hero>AnVIL hosts data from the following programs and also allows users to submit their own data to the platform.</hero>
-
-<data-ingestion-chart></data-ingestion-chart>
-
-## Current Consortia
-
-### [CCDG](https://www.genome.gov/Funded-Programs-Projects/NHGRI-Genome-Sequencing-Program/Centers-for-Common-Disease-Genomics) 
-
-The Centers for Common Disease Genomics are a collaborative large-scale genome sequencing effort to comprehensively identify rare risk and protective variants contributing to multiple common disease phenotypes.
-
-### [CMG](https://www.genome.gov/Funded-Programs-Projects/NHGRI-Genome-Sequencing-Program/Centers-for-Mendelian-Genomics-CMG)
-
-The Centers for Mendelian Genomics is a multi-center collaboration aimed at identifying the genes responsible for Mendelian phenotypes by whole exome and whole genome sequencing
-
-### [GTEx](https://gtexportal.org/home/)
-
-The Genotype-Tissue Expression (GTEx) project is an ongoing effort to build a comprehensive public resource to study tissue-specific gene expression and regulation. Samples were collected from 54 non-diseased tissue sites across nearly 1000 individuals, primarily for molecular assays including WGS, WES, and RNA-Seq.
-
-### [1000 G](https://www.internationalgenome.org/)
-
-The 1000 Genomes Project, launched in January 2008, is an international research effort to establish variation profiles across the human population. This open access data set continues to be a valuable resource to geneticists. 
-
-### [eMERGE](https://emerge-network.org/)
-
-The Electronic and MEdical Records and Genomics project (eMERGE) is a national network organized and funded by the NHGRI that combines DNA biorepositories with electronic medical record (EMR) systems for large scale, high-throughput genetic research in support of implementing genomic medicine.
-
-### [PAGE](https://www.genome.gov/Funded-Programs-Projects/Population-Architecture-Using-Genomics-and-Epidemiology)
-
-The Population Architecture Using Genomics and Epidemiology Consortium investigates ancestrally diverse populations to gain a better understanding of how genetic factors influence susceptibility to disease.
-
-### [HPRC](https://humanpangenome.org/)
-
-The Human Pangenome Reference Consortium aims to modernize the human reference to include a collection of diverse and highly accurate, haplotype-phased genome assemblies. This initiative will generate new technical standards in genome sequencing, scalable and reproducible assembly methods, and pangenomic tool development to ensure comprehensive variant discovery.
-
-## Planned Consortia
-
-The following consortia are planned for data ingestion. Additional consortia are under consideration and will be listed as they are approved.
-
-* [Covid19hg](https://www.covid19hg.org) - The COVID-19 host genetics initiative
-* [CSER](https://cser-consortium.org ) - Clinical Sequencing Evidence-Generating Research
-* [GTEx v9](https://gtexportal.org/home/) - Genotype-Tissue Expression Project
-* [NIA](https://www.nia.nih.gov/) - National Institute of Aging
-* [NIMH](https://nda.nih.gov/) - National Institute of Mental Health
-* [UDN](https://undiagnosed.hms.harvard.edu/) - Undiagnosed Disease Network
