+++ package.json
@@ -13,8 +13,6 @@
     "bowser": "^2.11.0",
     "classnames": "^2.3.1",
     "csv-parse": "^5.3.0",
-    "echarts": "^5.3.3",
-    "echarts-for-react": "^3.0.2",
     "gatsby": "^4.22.0",
     "gatsby-image": "^3.11.0",
     "gatsby-plugin-catch-links": "^4.22.0",
@@ -101,8 +99,7 @@
     "develop": "gatsby develop",
     "format": "prettier --write src/**/*.{js,ts,tsx}",
     "serve": "gatsby serve",
-    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\"",
-    "update-ingestion-chart": "cd scripts && node update-ingestion-chart.js"
+    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\""
   },
   "repository": {
     "type": "git",
