--- scripts/disease-codes/requirements.txt
@@ -1,36 +1,0 @@
-appnope==0.1.3
-asttokens==2.2.1
-backcall==0.2.0
-comm==0.1.2
-debugpy==1.6.6
-decorator==5.1.1
-executing==1.2.0
-importlib-metadata==6.0.0
-ipykernel==6.21.1
-ipython==8.9.0
-jedi==0.18.2
-jupyter_client==8.0.2
-jupyter_core==5.2.0
-matplotlib-inline==0.1.6
-nest-asyncio==1.5.6
-numpy==1.24.2
-packaging==23.0
-pandas==1.5.3
-parso==0.8.3
-pexpect==4.8.0
-pickleshare==0.7.5
-platformdirs==3.0.0
-prompt-toolkit==3.0.36
-psutil==5.9.4
-ptyprocess==0.7.0
-pure-eval==0.2.2
-Pygments==2.14.0
-python-dateutil==2.8.2
-pytz==2022.7.1
-pyzmq==25.0.0
-six==1.16.0
-stack-data==0.6.2
-tornado==6.2
-traitlets==5.9.0
-wcwidth==0.2.6
-zipp==3.12.1
