--- src/components/data-ingestion-chart/chart-data.d.ts
@@ -1,3 +1,0 @@
-export const startYear: number;
-export const endDateNum: number;
-export const monthDataByConsortium: Array<[string, number[]]>;
