--- src/components/data-ingestion-chart/chart-data.js
@@ -1,1 +1,0 @@
-module.exports = {startYear: 2019, endDateNum: 1680005415000, monthDataByConsortium: [["CMG",[0,0,0,0,0,0,17.128644391009267,0.7808175953490013,0,0.6128128666660014,0.05743039497699999,0,0,0,0,3.5756765721370085,0.5927409521590001,16.451395557817964,0.16892268232700003,0.08006962953900008,0,0,0,0,0,0,27.031150440698628,3.3774442480830524,7.699464402911038,23.63151326491386,0.32073578916900014,0.019889224874999993,23.460697592683523,4.661217395408002,0,0,0,0,0,0,0,0,0,0,0.37670465938722275,0,0,-0.02180983410142544,0,0,0,0.021809834101000004]],["CCDG",[0,0,0,0,0,0,869.1983943415781,0,0,0,9.749740747852716,13.001193225829223,141.26719459258484,3.0410524718900747,41.92611405959507,2.8009355321870086,99.55143361706448,68.49424698490543,4.703197115677608,3.142621276300036,6.641041000739863,9.311923897951008,19.22739880738797,1094.6330427724404,232.73173120888325,58.908836371442945,5.884423198021172,2.534865087829048,12.323597843710848,20.245349521416333,3.530279848263243,4.616442286606831,43.71549530220632,8.96475001579222,0.5228618078770039,0.6592883084929999,18.631897689590357,0.4455794882279919,0.14322222235900003,0.020727787115000007,1.3309445754989875,0,0,0,-12.181929454717967,0.000001400969,0,0,0,0.0005458331543884297,0,15.131314542348425]],["CSER",[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,103.3110451453889,0,0,0.38342475909200024,0,0,0,8.804050436140988,0,0,0,0,40.104537593147974,0.733655167322,0,0,0,-0.405800265633829,3.299197221208,0,-16.422364729178156,0.000009591994000000001,0.000005025098,9.41784e-7,0]],["PAGE",[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16.98311515566502,0,4.829581129616004,0,0,0,0.17819038051199998,0,0,0,0,0,0,0,0,0,0,0,0,0,2.3934792981191783e-14,0,0,-9.342027704117234e-16,0,0,0,0]],["Convergent Neuroscience",[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0.019527994124,5.304560467816027,0,0,0,0,0,0,0,0,0,1.3e-11,0,0,0.01006386121200003,1.4646454340159973,0,0,0,0,0.10688592220899708,0,0,0,0,3.5248812278943737,0,1.1123819310396432]],["GTEx",[0,0,0,0,0,0,48.22902503103397,179.01608274871387,0.049531960541,0,0.23568751218199852,2.831758323104997,0,0,0.002733772673,0,0,0,0,0,0,0,0,0,0,0.000002505733,16.029294633145998,0,0,0,0.0005807290040000001,2.0765244206279987,0,0,0.06831088900299999,47.85165542406105,0.04946172996000002,0,3.0799709758719995,9.72015487450901,0.11613533507999975,0,0,0,15.419249841536777,0.026782469113,0,-13.313955028889444,0,0,0,13.287101724916992]],["eMERGE",[0,0,0,0,0,0,0,0,0,0,0.03845468172099999,35.41767834632157,10.350965019381492,27.4979590912533,0,0,0.0009907395299999999,0,0,0,0,0,0,0,0,0,0.122195809945,0,0,0,0,0,0,0,0,0,0,0,0,0.003817724122,0,0,0,0,-2.06410543159216e-10,0.0680352463460001,0.11883566238210533,0.3200412769309993,0.6556088854050012,0,0,0]],["CMH",[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,39.824940143925026,0,0,0,0,0,0,0,0,57.73523027949878,0,0,0,0,0,0,1.0658141036401503e-14]],["HPRC",[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2.002047553691,6.480807057602995,5.4239798493749944,6.988485520853159,88.93271939117638,9.027225391019,0.1643150849199999,9.624987752487995,8.272057081780067,0.238920883717,13.443224799058976,8.556774350035994,15.10451636443495,20.97689797706216,0,19.688462734222963,8.545786496501997,9.362255559531008,16.95935052478799,0.502142550307,0.1091190128640023,0,0,0,0,10.300859394361044,0,0,15.049179423123672,19.36360239379092,111.87808171916711,46.32050368568021,10.669905924781288]],["T2T",[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0.010242241995999999,110.67035206444598,0.022589064584,9.519774953765003,236.1742525229733,56.75229338107303,12.119580304264998,0,0,0.013151222734999999,0.44501076257999994,146.37376617182653,0,0,0,0,-2.0375745492232284e-11,0,0,-2.5801540063014917e-14,0,0,0,0]],["1000G",[0,0,41.24746970906762,0.04554811050600001,0,0,0,0,1.4205709577800065,0,0,0,15.166951096314506,0,15.097331035121826,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-6.3105071812770745e-12,0,0,0,0,0,0,1.4138445262517152]],["WGSPD1",[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,145.63215896197332,0,0,31.730904564461284,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-3.3532316374440473e-12,0,0,0,0,0,0,1.2695573758936263e-14]],["Unspecified",[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0.5984915621190003,0,0,-0.5984915621190003,0,0,0,0]]]};
