+++ src/components/home/section-dataset-cloud/section-dataset-cloud.tsx
@@ -9,21 +9,21 @@
 import React, { FC } from "react";
 
 // App dependencies
+import { DatasetCloudStaticQuery } from "../../../hooks/dataset-cloud-query";
+import { getDatasetsEnvironmentUrl } from "../../../utils/environment/environment.service";
+import ButtonCta from "../../button-cta/button-cta";
 import ButtonSize from "../../button/button-size";
 import ButtonTheme from "../../button/button-theme";
-import ButtonCta from "../../button-cta/button-cta";
 import { ICard } from "../../card/card";
-import DatasetCard from "../dataset-card/dataset-card";
-import { DatasetCloudStaticQuery } from "../../../hooks/dataset-cloud-query";
-import Section from "../section/section";
-import SectionContent from "../section/section-content/section-content";
 import Stats from "../../stats/stats";
 import { Target } from "../../target/target.model";
-import { getDatasetsEnvironmentUrl } from "../../../utils/environment/environment.service";
+import DatasetCard from "../dataset-card/dataset-card";
+import Section from "../section/section";
+import SectionContent from "../section/section-content/section-content";
 
 // Styles
-import { datasetStats, sectionCards } from "./section-dataset-cloud.module.css";
 import { sectionHero } from "../section-workspace-cloud/section-workspace-cloud.module.css";
+import { datasetStats, sectionCards } from "./section-dataset-cloud.module.css";
 
 const SectionDatasetCloud: FC = (): JSX.Element => {
   const frontmatter = DatasetCloudStaticQuery();
